Changed enums to CAPSLOCK to adhere to TDD specifications
Resolved warnings of APIENTRY

Note: To confirm with Elie if pragma warning can be used to suppress warnings caused from GLM type_quat.hpp

+++ Source/Systems/FrameRateController.cpp
@@ -117,6 +117,6 @@ int FrameRateController::GetFPS()
 		newFPS = GetFrames() - prevFPS;
 		prevFPS = GetFrames();
 	}
-	seconds = TimeElapsed(s);
+	seconds = static_cast<int>(TimeElapsed(s));
 	return newFPS;
 }
